+++ src/Symfony/Component/Security/Http/Event/DeauthenticatedEvent.php
@@ -15,7 +15,11 @@ use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
 use Symfony\Contracts\EventDispatcher\Event;
 
 /**
- * Deauthentication happens in case the user has changed when trying to refresh the token.
+ * Deauthentication happens in case the user has changed when trying to
+ * refresh the token.
+ *
+ * Use {@see TokenDeauthenticatedEvent} if you want to cover all cases where
+ * a session is deauthenticated.
  *
  * @author Hamza Amrouche <hamza.simperfit@gmail.com>
  */