+++ src/Symfony/Component/Security/Http/Tests/LoginLink/LoginLinkHandlerTest.php
@@ -13,17 +13,20 @@ namespace Symfony\Component\Security\Http\Tests\LoginLink;
 
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
+use Psr\Cache\CacheItemPoolInterface;
+use Symfony\Component\Cache\Adapter\ArrayAdapter;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\PropertyAccess\PropertyAccess;
 use Symfony\Component\PropertyAccess\PropertyAccessorInterface;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use Symfony\Component\Routing\RequestContext;
 use Symfony\Component\Security\Core\Exception\UserNotFoundException;
+use Symfony\Component\Security\Core\Signature\ExpiredSignatureStorage;
+use Symfony\Component\Security\Core\Signature\SignatureHasher;
 use Symfony\Component\Security\Core\User\UserInterface;
 use Symfony\Component\Security\Core\User\UserProviderInterface;
 use Symfony\Component\Security\Http\LoginLink\Exception\ExpiredLoginLinkException;
 use Symfony\Component\Security\Http\LoginLink\Exception\InvalidLoginLinkException;
-use Symfony\Component\Security\Http\LoginLink\ExpiredLoginLinkStorage;
 use Symfony\Component\Security\Http\LoginLink\LoginLinkHandler;
 
 class LoginLinkHandlerTest extends TestCase
@@ -34,15 +37,18 @@ class LoginLinkHandlerTest extends TestCase
     private $userProvider;
     /** @var PropertyAccessorInterface */
     private $propertyAccessor;
-    /** @var MockObject|ExpiredLoginLinkStorage */
+    /** @var MockObject|ExpiredSignatureStorage */
     private $expiredLinkStorage;
+    /** @var CacheItemPoolInterface */
+    private $expiredLinkCache;
 
     protected function setUp(): void
     {
         $this->router = $this->createMock(UrlGeneratorInterface::class);
         $this->userProvider = new TestLoginLinkHandlerUserProvider();
         $this->propertyAccessor = PropertyAccess::createPropertyAccessor();
-        $this->expiredLinkStorage = $this->createMock(ExpiredLoginLinkStorage::class);
+        $this->expiredLinkCache = new ArrayAdapter();
+        $this->expiredLinkStorage = new ExpiredSignatureStorage($this->expiredLinkCache, 360);
     }
 
     /**
@@ -118,13 +124,12 @@ class LoginLinkHandlerTest extends TestCase
         $user = new TestLoginLinkHandlerUser('weaverryan', 'ryan@symfonycasts.com', 'pwhash');
         $this->userProvider->createUser($user);
 
-        $this->expiredLinkStorage->expects($this->once())
-            ->method('incrementUsages')
-            ->with($signature);
-
         $linker = $this->createLinker(['max_uses' => 3]);
         $actualUser = $linker->consumeLoginLink($request);
         $this->assertEquals($user, $actualUser);
+
+        $item = $this->expiredLinkCache->getItem(rawurlencode($signature));
+        $this->assertSame(1, $item->get());
     }
 
     public function testConsumeLoginLinkWithExpired()
@@ -172,10 +177,9 @@ class LoginLinkHandlerTest extends TestCase
         $user = new TestLoginLinkHandlerUser('weaverryan', 'ryan@symfonycasts.com', 'pwhash');
         $this->userProvider->createUser($user);
 
-        $this->expiredLinkStorage->expects($this->once())
-            ->method('countUsages')
-            ->with($signature)
-            ->willReturn(3);
+        $item = $this->expiredLinkCache->getItem(rawurlencode($signature));
+        $item->set(3);
+        $this->expiredLinkCache->save($item);
 
         $linker = $this->createLinker(['max_uses' => 3]);
         $linker->consumeLoginLink($request);
@@ -199,7 +203,7 @@ class LoginLinkHandlerTest extends TestCase
             'route_name' => 'app_check_login_link_route',
         ], $options);
 
-        return new LoginLinkHandler($this->router, $this->userProvider, $this->propertyAccessor, $extraProperties, 's3cret', $options, $this->expiredLinkStorage);
+        return new LoginLinkHandler($this->router, $this->userProvider, new SignatureHasher($this->propertyAccessor, $extraProperties, 's3cret', $this->expiredLinkStorage, $options['max_uses'] ?? null), $options);
     }
 }
 
@@ -209,7 +213,7 @@ class TestLoginLinkHandlerUserProvider implements UserProviderInterface
 
     public function createUser(TestLoginLinkHandlerUser $user): void
     {
-        $this->users[$user->getUsername()] = $user;
+        $this->users[$user->getUserIdentifier()] = $user;
     }
 
     public function loadUserByIdentifier(string $userIdentifier): TestLoginLinkHandlerUser